requirements.txt
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.7
black==21.5b0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
Django==3.2
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
django-simple-history==3.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
flake8==3.9.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
pathspec==0.8.1
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
yapf==0.31.0
